refs #5111 - accounting for `value` potentially being `false`
There were some occurrences where `value` was `false` which were causing
the `||` to fall through and the method returning `''` (empty string).
This also alleviates having to use an `||` at all because even if
`value` turns out to be `nil`, `nil.to_s` evaluates to `''`.

Now that `#formatted_value` is always going to return some kind of
string, we don't need to worry about nilness inside of `#self.values`.